# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AnimusForge API
  description: |
    ## Adaptive AI Persona Framework with Ethical Governance

    AnimusForge is an enterprise-grade framework for creating, managing, and evolving
    AI personas with built-in ethical governance, kill-switch mechanisms, and
    comprehensive memory systems.

    ### Key Features

    - **Persona Management**: Create, update, activate, and evolve AI personas with
      configurable traits, cognitive patterns, and behavioral constraints.
    - **LLM Gateway**: Multi-provider LLM access with automatic failover, load balancing,
      and cost optimization across OpenAI, Anthropic, DeepSeek, and GLM.
    - **Kill-Switch**: Emergency stop and recovery mechanisms for immediate persona
      deactivation with audit trail and recovery workflows.
    - **Memory System**: Vector-based semantic memory and graph-based knowledge storage
      for persistent persona context and learning.
    - **Ethics Engine**: Real-time content evaluation, audit logging, and compliance
      reporting across multiple ethical categories.
    - **System Monitoring**: Health checks, Prometheus metrics, and configuration
      management endpoints.

    ### Authentication

    JWT authentication is supported. Include `Authorization: Bearer <token>` header
    for authenticated requests. Some endpoints may require elevated permissions.

    ### Rate Limiting

    All endpoints are rate-limited. Check response headers:
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-03-06T12:00:00Z"
...
    LLMProvider:
      type: string
      enum:
        - openai
        - anthropic
        - deepseek
        - glm
        - local
        - custom
    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant, function]
        content:
          type: string
        name:
          type: string
    GenerationRequest:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
          minItems: 1
        model:
          type: string
          example: "gpt-4-turbo"
        provider:
          $ref: "#/components/schemas/LLMProvider"
        persona_id:
          $ref: "#/components/schemas/UUID"
        max_tokens:
          type: integer
          default: 1024
          maximum: 128000
        temperature:
          type: number
          format: float
          default: 0.7
        stream:
          type: boolean
          default: false
    TokenUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    GenerationResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        content:
          type: string
        model:
          type: string
        provider:
          $ref: "#/components/schemas/LLMProvider"
        finish_reason:
          type: string
          enum: [stop, length, content_filter, error, cancelled]
        usage:
          $ref: "#/components/schemas/TokenUsage"
        latency_ms:
          type: number
          format: float
        created_at:
          $ref: "#/components/schemas/Timestamp"

    # ==================== Kill-Switch Schemas ====================
...
paths:
...
  /api/v1/llm/generate:
    post:
      tags:
        - LLM Gateway
      summary: Generate text
      description: Generate text using the LLM gateway with automatic provider selection.
      operationId: generateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationRequest"
            example:
              messages:
                - role: "user"
                  content: "Explain quantum computing in simple terms"
              model: "gpt-4-turbo"
              max_tokens: 1024
              temperature: 0.7
      responses:
        "200":
          description: Text generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "503":
          description: No providers available
...
  /api/v1/llm/providers/{provider_name}/status:
    get:
      tags:
        - LLM Gateway
      summary: Get provider status
      description: Retrieve the status and health of a specific LLM provider.
      operationId: getProviderStatus
      parameters:
        - name: provider_name
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LLMProvider"
      responses:
        "200":
          description: Provider status
        "404":
          description: Provider not found
  /api/v1/llm/providers/{provider_name}/enable:
    post:
      tags:
        - LLM Gateway
      summary: Enable provider
      description: Enable an LLM provider for use.
      operationId: enableProvider
      parameters:
        - name: provider_name
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LLMProvider"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key:
                  type: string
                default_model:
                  type: string
      responses:
        "200":
          description: Provider enabled successfully
        "404":
          description: Provider not found
  /api/v1/llm/providers/{provider_name}/disable:
    post:
      tags:
        - LLM Gateway
      summary: Disable provider
      description: Disable an LLM provider.
      operationId: disableProvider
      parameters:
        - name: provider_name
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LLMProvider"
      responses:
        "200":
          description: Provider disabled successfully
        "404":
          description: Provider not found

  # ==================== Kill-Switch ====================
...